% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doGSEA.R
\name{addGeneAbundance}
\alias{addGeneAbundance}
\title{addGeneAbundance}
\usage{
addGeneAbundance(
  stana,
  candSp,
  IDs,
  target = "KO",
  how = sum,
  newCol = "gene",
  disc = NULL,
  discNumeric = TRUE
)
}
\arguments{
\item{stana}{stana object}

\item{candSp}{candidate species}

\item{IDs}{gene IDs to add}

\item{target}{KO or genes}

\item{how}{how to combine multiple IDs}

\item{newCol}{new column name}

\item{disc}{discretize the abundance by the threshold. function for calculating
threshold, like {median}}

\item{discNumeric}{convert discrete value to numeric}
}
\value{
stana
}
\description{
Add the specified gene copy number to metadata
}
