% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alleleStat.R
\name{alleleStat}
\alias{alleleStat}
\title{alleleStat}
\usage{
alleleStat(stana, sp, cl, base = "maj", deleteZeroDepth = FALSE)
}
\arguments{
\item{stana}{stana object}

\item{sp}{candidate species}

\item{cl}{named list of clusters of samples}

\item{base}{"maj","ref", or "majref"}

\item{deleteZeroDepth}{delete snvs with zero depth}
}
\description{
Output statistics of MAF, using site-by-sample 
minor allele frequency table output by MIDAS1 and MIDAS2.
}
