% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annot.R
\name{anno_eggNOG_keywords}
\alias{anno_eggNOG_keywords}
\title{anno_eggNOG_keywords}
\usage{
anno_eggNOG_keywords(
  split,
  genes,
  tib,
  removeHigh = TRUE,
  removeAdditional = NULL,
  argList = list()
)
}
\arguments{
\item{split}{named list of cluster}

\item{genes}{gene names}

\item{tib}{result of checkEGGNOG()}

\item{removeHigh}{remove frequent words}

\item{removeAdditional}{remove these words
passed to anno_word_cloud}

\item{argList}{will be passed to anno_word_cloud}
}
\description{
Annotate the heatmap using functional terms derived from eggNOG.
The function is derived and modified from the example of 
KeywordsEnrichment (https://github.com/jokergoo/KeywordsEnrichment) 
and using simplifyEnrichment package.
Can be replaced by providing simplyfyEnrichment::anno_word_cloud()
}
