% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annot.R
\name{anno_PATRIC_keywords}
\alias{anno_PATRIC_keywords}
\title{anno_PATRIC_keywords}
\usage{
anno_PATRIC_keywords(
  split,
  genes,
  fnc = "pathway_name",
  removeHigh = TRUE,
  removeAdditional = NULL,
  argList = list()
)
}
\arguments{
\item{split}{named list of cluster}

\item{genes}{gene names}

\item{fnc}{"pathway_name" or "ec_description"}

\item{removeHigh}{remove frequent words}

\item{removeAdditional}{remove these words
passed to anno_word_cloud}

\item{argList}{will be passed to anno_word_cloud}
}
\description{
Annotate the heatmap using functional terms derived from PATRIC.
The function is derived and modified from the example of 
KeywordsEnrichment (https://github.com/jokergoo/KeywordsEnrichment) 
and using simplifyEnrichment package.
Can be replaced by providing simplyfyEnrichment::anno_word_cloud() 
with split and results from checkPATRICSimple().
}
