% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doGSEA.R
\name{calcGF}
\alias{calcGF}
\title{calcGF}
\usage{
calcGF(stana, candSp = NULL, how = sum, annot = "eggNOG", column = "KEGG_ko")
}
\arguments{
\item{stana}{stana object}

\item{candSp}{candidate species ID}

\item{how}{how to summarize multiple gene CN assigned to the same KO}

\item{annot}{eggNOG or manual}

\item{column}{When eggNOG, which family to summarize, default to KEGG_ko}
}
\description{
calcGF
}
