% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doGSEA.R
\name{calcKO}
\alias{calcKO}
\title{calcKO}
\usage{
calcKO(stana, candSp = NULL, how = sum, annot = "eggNOG")
}
\arguments{
\item{stana}{stana object}

\item{candSp}{candidate species ID}

\item{how}{how to summarize multiple gene CN assigned to the same KO}

\item{annot}{eggNOG or manual}
}
\description{
calcKO
}
