% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stana.R
\name{changeColors}
\alias{changeColors}
\title{changeColors}
\usage{
changeColors(stana, colors)
}
\arguments{
\item{stana}{stana object}

\item{colors}{color vector}
}
\value{
stana object
}
\description{
changeColors
}
