% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stana.R
\name{check,stana-method}
\alias{check,stana-method}
\title{check
check and output statistics based on conditional formulas for summary}
\usage{
\S4method{check}{stana}(x, exp, target = "snps")
}
\arguments{
\item{x}{stana boject}

\item{exp}{expression for filtering the snps summary}

\item{target}{snps or genes}
}
\description{
check
check and output statistics based on conditional formulas for summary
}
