% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stana.R
\name{check}
\alias{check}
\title{check
check and output statistics based on conditional formulas}
\usage{
check(x, exp, target = "snps")
}
\description{
check
check and output statistics based on conditional formulas
}
