% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkPATRIC.R
\name{checkPATRIC}
\alias{checkPATRIC}
\title{checkPATRIC}
\usage{
checkPATRIC(
  genes,
  whichToCount = "ec_description",
  delSize = 0,
  showSubset = 5,
  colText = "cat",
  rel = FALSE,
  lyt = "nicely",
  nodeSize = "count",
  skipGraph = FALSE
)
}
\arguments{
\item{genes}{named list of genes}

\item{whichToCount}{ec_description, ec_number, pathway_name, pathway_id}

\item{delSize}{functions with count above this threshold is included}

\item{showSubset}{in graph, show this number of text ordered by frequency (default: 5)}

\item{colText}{how to color text (default: cat, EC and KEGG)}

\item{rel}{use relative frequency}

\item{lyt}{ggraph layout (default: nicely)}

\item{nodeSize}{'count' or 'degree'}

\item{skipGraph}{skip graph plotting}
}
\description{
Obtain gene information from PATRIC server.
Input named list of genes, and returns queried results,
Count for functionality after removing duplicate entries,
and network representation of EC and KEGG pathway.
For midas_db_v1.2 only.
}
