% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkPATRICSimple.R
\name{checkPATRICSimple}
\alias{checkPATRICSimple}
\title{checkPATRICSimple}
\usage{
checkPATRICSimple(genes, whichToCount = "ec_description")
}
\arguments{
\item{genes}{named list of genes}

\item{whichToCount}{ec_description, ec_number, pathway_name, pathway_id}
}
\description{
Obtain gene information from PATRIC server.
Input named list of genes, and returns named list of functions.
For midas_db_v1.2 only.
}
