% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadProfile.R
\name{checkProfile}
\alias{checkProfile}
\title{checkProfile}
\usage{
checkProfile(midas_merge_dir, cl, filtNum = 2)
}
\arguments{
\item{midas_merge_dir}{output directory of merge_midas.py}

\item{cl}{named list of sample IDs}

\item{filtNum}{The species with number above this threshold
for each category is returned}
}
\description{
Assess profile for species and return filtered species 
based on the number of samples for each category.
For MIDAS only. Deprecated, use load* functions instead.
}
