% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cnDiscretize}
\alias{cnDiscretize}
\title{Discretize copy number matrix at specified cutoff}
\usage{
cnDiscretize(stana, species, cutoff = 0.35)
}
\arguments{
\item{stana}{stana object}

\item{species}{species id}

\item{cutoff}{cutoff value, default to 0.35}
}
\description{
Discretize copy number matrix at specified cutoff
}
