% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineGenes.R
\name{combineGenes}
\alias{combineGenes}
\title{combineGenes}
\usage{
combineGenes(stana_list, species, argList = list(), output_seq = FALSE)
}
\arguments{
\item{stana_list}{stana list}

\item{species}{species ID}
}
\value{
new stana object
}
\description{
(experimental) use common genes across multiple datasets to
merge gene copy numbers
}
