% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertWideToMaf.R
\name{combineMaf}
\alias{combineMaf}
\title{combineMaf
combine the maf table produced by convertWideToMaf()}
\usage{
combineMaf(mafs)
}
\arguments{
\item{mafs}{named list of the results of convertWideToMaf()}
}
\value{
stana object
}
\description{
combineMaf
combine the maf table produced by convertWideToMaf()
}
