% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareGenes.R
\name{compareGenes}
\alias{compareGenes}
\title{compareGenes}
\usage{
compareGenes(
  stana,
  species = NULL,
  geneID = NULL,
  cl = NULL,
  argList = list(),
  verbose_zero = FALSE
)
}
\arguments{
\item{stana}{stana object}

\item{species}{candidate species}

\item{geneID}{gene ID to be plotted}

\item{cl}{named list of clusters}

\item{argList}{argument list for wilcox.exact}

\item{verbose_zero}{show zero abundance genes}
}
\value{
ggplot
}
\description{
Compare the gene abundances by recursively performing Wilcoxon rank-sum tests
Use exactRankTests package, and multiple comparison adjustment is not performed
}
