% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensusSeq.R
\name{consensusSeq}
\alias{consensusSeq}
\title{consensusSeq}
\usage{
consensusSeq(target = "MIDAS1", ...)
}
\arguments{
\item{midas_merge_dir}{merge directory}

\item{species}{species vectors}
}
\description{
Output consensus sequences from merged SNV output.
Optionally, return phylogenetic tree inferred by `phangorn`.
If specified cluster of samples, additionally returns plot by `ggtree`.
}
