% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensusSeq.R
\name{consensusSeqMIDAS1}
\alias{consensusSeqMIDAS1}
\title{consensusSeqMIDAS1}
\usage{
consensusSeqMIDAS1(
  stana,
  species,
  mean_depth = 0,
  fract_cov = 0,
  site_depth = 5,
  site_ratio = 5,
  site_maf = 0.01,
  allele_support = 0.5,
  site_prev = 0.9,
  cl = NULL,
  max_sites = Inf,
  tree = FALSE,
  max_samples = Inf,
  keep_samples = NULL,
  exclude_samples = NULL,
  rand_samples = NULL
)
}
\arguments{
\item{stana}{stana obj}

\item{species}{candidate species vector}

\item{mean_depth}{parameter for filtering}

\item{fract_cov}{parameter for filtering}

\item{site_depth}{parameter for filtering}

\item{site_ratio}{parameter for filtering}

\item{site_maf}{parameter for filtering}

\item{allele_support}{parameter for filtering}

\item{site_prev}{parameter for filtering}

\item{cl}{cluster, if plot cladogram}

\item{max_sites}{currently not implemented}

\item{tree}{if perform tree inference using dist.ml()
and NJ() in default parameters}

\item{max_samples}{currently not implemented}

\item{keep_samples}{currently not implemented}

\item{exclude_samples}{currently not implemented}

\item{rand_samples}{currently not implemented}
}
\description{
consensusSeqMIDAS1
}
