% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensusSeqFast.R
\name{consensusSeqMIDAS2}
\alias{consensusSeqMIDAS2}
\title{consensusSeqMIDAS2Fast}
\usage{
consensusSeqMIDAS2(
  stana,
  species = NULL,
  output_seq = FALSE,
  mean_depth = 0,
  fract_cov = 0,
  site_depth = 5,
  site_ratio = 5,
  site_maf = 0.01,
  allele_support = 0.5,
  site_prev = 0.9,
  locus_type = "CDS",
  site_list = NULL,
  cl = NULL,
  max_sites = Inf,
  tree = FALSE,
  max_samples = Inf,
  keep_samples = NULL,
  exclude_samples = NULL,
  rand_samples = NULL,
  return_mat = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{stana}{stana obj}

\item{species}{candidate species vector}

\item{output_seq}{whether to output actual FASTA file}

\item{mean_depth}{parameter for filtering}

\item{fract_cov}{parameter for filtering}

\item{site_depth}{parameter for filtering}

\item{site_ratio}{parameter for filtering}

\item{site_maf}{parameter for filtering}

\item{allele_support}{parameter for filtering}

\item{site_prev}{parameter for filtering}

\item{locus_type}{locus type to be included, default to CDS}

\item{site_list}{site list to be included}

\item{cl}{cluster, if plot cladogram}

\item{max_sites}{default to Inf}

\item{tree}{if perform tree inference}

\item{max_samples}{currently not implemented}

\item{keep_samples}{currently not implemented}

\item{exclude_samples}{currently not implemented}

\item{rand_samples}{currently not implemented}

\item{return_mat}{return matrix of characters}

\item{verbose}{output current status}
}
\description{
consensusSeqMIDAS2Fast
}
