% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensusSeqRaw.R
\name{consensusSeqRaw}
\alias{consensusSeqRaw}
\title{consensusSeqRaw}
\usage{
consensusSeqRaw(target = "MIDAS1", midas_merge_dir, species, ...)
}
\arguments{
\item{target}{MIDAS1 or MIDAS2}

\item{midas_merge_dir}{merge directory}

\item{species}{species vectors}

\item{...}{filters, passed to corresponding functions}
}
\description{
Output consensus sequences from merged SNV output.
Optionally, return phylogenetic tree inferred by `phangorn`.
If specified cluster of samples, additionally returns plot by `ggtree`.
}
