% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertWideToMaf.R
\name{convertWideToMaf}
\alias{convertWideToMaf}
\title{convertWideToMaf}
\usage{
convertWideToMaf(df)
}
\arguments{
\item{df}{data frame of six columns}
}
\value{
species-named list of MAF data.frame
}
\description{
should be six column layout with the column corresponding to
"species_id","snv_id","ref_allele","alt_allele","ref_count","alt_count"
}
