% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doAdonis.R
\name{doAdonis}
\alias{doAdonis}
\title{doAdonis}
\usage{
doAdonis(
  stana,
  specs,
  cl = NULL,
  target = "snps",
  formula = NULL,
  distMethod = "manhattan",
  pcoa = FALSE,
  AAfunc = dist.ml,
  AAargs = list(),
  maj = FALSE,
  deleteZeroDepth = FALSE,
  useMeta = FALSE,
  argList = list(),
  distArg = list()
)
}
\arguments{
\item{stana}{stana object}

\item{specs}{species to be examined}

\item{cl}{named list of samples}

\item{target}{tree, snps, genes, fasta}

\item{formula}{pass to adonis2, specify distance matrix as d.}

\item{distMethod}{distance method passed to dist() (default, manhattan)}

\item{pcoa}{if TRUE, performs ape::pcoa on dist and plot them}

\item{AAfunc}{if choose `fasta`, provide function for calculating distance}

\item{AAargs}{provided to `AAfunc`}

\item{maj}{major allele distance}

\item{deleteZeroDepth}{delete zero depth snvs (denoted as `-1`)
Otherwise the cell is treated as NA}

\item{useMeta}{use metadata slot, not cl slot}

\item{argList}{parameters passed to adonis2}

\item{distArg}{passed to `dist()`}
}
\description{
Perform PERMANOVA on distance matrix based 
on various distances based on such as SNV frequency or 
gene matrix using `adonis2`.
Named list of samples are to be provided.
Note that this function currently performs 
comparison using one category.
for complex design, one should do it manually.
}
