% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doAdonisMIDAS1Raw.R
\name{doAdonisMIDAS1Raw}
\alias{doAdonisMIDAS1Raw}
\title{doAdonisMIDAS1Raw}
\usage{
doAdonisMIDAS1Raw(
  midas_merge_dir,
  specs,
  cl,
  target = "snps",
  formula = NULL,
  distMethod = "manhattan",
  ...
)
}
\arguments{
\item{midas_merge_dir}{output directory of midas2 merge}

\item{specs}{species to be examined}

\item{cl}{named list of samples}

\item{target}{snps, presabs, or copynum}

\item{formula}{pass to adonis2, specify distance matrix as d.}

\item{distMethod}{distance method passed to dist() (default, manhattan)}

\item{...}{parameters passed to adonis2}
}
\description{
Perform PERMANOVA on distance matrix based on SNV frequency.
Named list of samples are to be provided.
Note that this function performs comparison using one category.
for complex design, one should do it manually.
}
