% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doAdonisMIDAS2.R
\name{doAdonisMIDAS2}
\alias{doAdonisMIDAS2}
\title{doAdonisMIDAS2}
\usage{
doAdonisMIDAS2(
  midas_merge_dir,
  specs,
  cl,
  target = "snps",
  taxtbl = NULL,
  formula = NULL,
  distMethod = "manhattan",
  ...
)
}
\arguments{
\item{midas_merge_dir}{output directory of midas2 merge}

\item{specs}{species to be examined}

\item{cl}{named list of samples}

\item{taxtbl}{data.frame of 6-digit MIDAS species IDs 
as row.names, and have `GTDB species` column}

\item{formula}{pass to adonis2, specify distance matrix as d.}

\item{...}{parameters passed to adonis2}

\item{distMethods}{distance method passed to dist() (default, manhattan)}
}
\description{
Perform PERMANOVA on distance matrix based on SNV frequency.
Named list of samples are to be provided.
Note that this function performs comparison using one category.
for complex design, one should do it manually.
The location of `lz4` binary must be in PATH.
}
