% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doBoruta.R
\name{doBoruta}
\alias{doBoruta}
\title{doBoruta}
\usage{
doBoruta(
  stana,
  sp,
  cl = NULL,
  doFix = TRUE,
  target = "genes",
  mat = NULL,
  whichToCount = "ec_description",
  argList = list(),
  deleteZeroDepth = FALSE
)
}
\arguments{
\item{stana}{stana object}

\item{sp}{candidate species ID}

\item{cl}{named list of category}

\item{doFix}{performs TentativeRoughFix on Boruta result}

\item{target}{default to snps}

\item{mat}{if target is not snps, provide preprocessed gene matrix
otherwise the raw gene matrix is used.}

\item{whichToCount}{which to show on box plots,
pass to checkPATRIC() (in MIDAS1)}

\item{argList}{passed to Boruta()}

\item{deleteZeroDepth}{delete zero depth SNV}
}
\value{
list of Boruta object and plot
}
\description{
feature selection of classification between category
using snv frequency and gene matrix
}
