% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doL0.R
\name{doL0}
\alias{doL0}
\title{doL0}
\usage{
doL0(
  stana,
  species,
  cl = NULL,
  doFix = TRUE,
  target = "genes",
  mat = NULL,
  whichToCount = "ec_description",
  argList = list(),
  deleteZeroDepth = FALSE
)
}
\arguments{
\item{stana}{stana object}

\item{species}{candidate species ID}

\item{cl}{named list of category}

\item{target}{default to snps}

\item{mat}{if target is not snps, provide preprocessed gene matrix
otherwise the raw gene matrix is used.}

\item{argList}{passed to L0Learn()}

\item{deleteZeroDepth}{delete zero depth SNV}
}
\value{
L0Learn object
}
\description{
feature selection of classification between category
using snv frequency and gene matrix, based on L0Learn
}
