% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkEGGNOG.R
\name{drawEGGNOG}
\alias{drawEGGNOG}
\title{drawEGGNOG}
\usage{
drawEGGNOG(annot_file, geneIDs, candPlot)
}
\arguments{
\item{annot_file}{annotation file}

\item{geneIDs}{geneIDs}

\item{candPlot}{vector consisting of 
"COG_category" "Description" "Preferred_name" "GOs"           
"EC"             "KEGG_ko"        "KEGG_Pathway"   "KEGG_Module"    "KEGG_Reaction" 
"KEGG_rclass"    "BRITE"          "KEGG_TC"        "CAZy"           "BiGG_Reaction" 
"PFAMs"}
}
\value{
list containing ggplot2 object and graph
}
\description{
draw the graph representing the eggNOG annotation
}
