% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stana.R
\name{filter,stana-method}
\alias{filter,stana-method}
\title{filter
filter the stana object based on species ID}
\usage{
\S4method{filter}{stana}(x, ids, target = "snps")
}
\arguments{
\item{x}{stana object}

\item{ids}{species ID}

\item{target}{snps or genes}
}
\description{
filter
filter the stana object based on species ID
}
