% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stana.R
\name{fromHumann}
\alias{fromHumann}
\title{fromHumann
import from HUMAnN profiles}
\usage{
fromHumann(df)
}
\arguments{
\item{df}{data frame (row.names as taxonomy and column names as sample names)}
}
\value{
stana object
}
\description{
fromHumann
import from HUMAnN profiles
}
