% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InStrainFuncs.R
\name{genomeHeatmap}
\alias{genomeHeatmap}
\title{genomeHeatmap}
\usage{
genomeHeatmap(
  stana,
  species,
  column = "conANI",
  cl = NULL,
  heatmapArgs = list()
)
}
\arguments{
\item{stana}{stana object}

\item{species}{interesting species}

\item{column}{column to plot, default to conANI}

\item{cl}{grouping, if NULL, it automatically obtain grouping from stana object}

\item{heatmapArgs}{list of parameters passed to Heatmap()}
}
\value{
Heatmap by ComplexHeatmap
}
\description{
Draw a heatmap representing columns described in genome-wide
comparison table.
}
