% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stana.R
\name{getAdonis,stana-method}
\alias{getAdonis,stana-method}
\title{getAdonis
get adonis list from stana}
\usage{
\S4method{getAdonis}{stana}(x)
}
\arguments{
\item{x}{stana object}
}
\description{
getAdonis
get adonis list from stana
}
