% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stana.R
\name{getCl,stana-method}
\alias{getCl,stana-method}
\title{getCl
get grouping information list from stana}
\usage{
\S4method{getCl}{stana}(x)
}
\arguments{
\item{x}{stana object}
}
\description{
getCl
get grouping information list from stana
}
