% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stana.R
\name{getFasta,stana-method}
\alias{getFasta,stana-method}
\title{getFasta
get fasta list from stana}
\usage{
\S4method{getFasta}{stana}(x)
}
\arguments{
\item{x}{stana object}
}
\description{
getFasta
get fasta list from stana
}
