% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stana.R
\name{getGeneID,stana-method}
\alias{getGeneID,stana-method}
\title{getGeneID
get gene IDs from stana}
\usage{
\S4method{getGeneID}{stana}(x, candSp)
}
\arguments{
\item{x}{stana object}

\item{candSp}{species}
}
\description{
getGeneID
get gene IDs from stana
}
