% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stana.R
\name{getGeneID}
\alias{getGeneID}
\title{getGeneID
get gene IDs from stana}
\usage{
getGeneID(x, candSp)
}
\arguments{
\item{x}{stana object}

\item{candSp}{species}
}
\description{
getGeneID
get gene IDs from stana
}
