% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGenes.R
\name{getGenes}
\alias{getGenes}
\title{getGenes}
\usage{
getGenes(
  midas_merge_dir,
  candidate = "all",
  pa = "presabs",
  km = 20,
  filUp = 1,
  filDown = 0,
  heatmap = FALSE,
  seed = 1
)
}
\arguments{
\item{midas_merge_dir}{output directory of merge_midas.py}

\item{candidate}{candidate species, default to all}

\item{pa}{"presabs" or "copynum"}

\item{km}{kmeans number if perform heatmap clustering}

\item{filUp}{filter parameter for upper limit}

\item{filDown}{filter parameter for lower limit}

\item{heatmap}{whether to draw heatmap}

\item{seed}{heatmap seed}
}
\description{
Obtain gene matrix from midas merge directory from MIDAS.
Additionally, heatmap can be stored with clustering information.
}
