% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stana.R
\name{getKOID}
\alias{getKOID}
\title{getKOID
get gene family IDs from stana}
\usage{
getKOID(x, candSp)
}
\arguments{
\item{x}{stana object}

\item{candSp}{species}
}
\description{
getKOID
get gene family IDs from stana
}
