% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stana.R
\name{getSlot,stana-method}
\alias{getSlot,stana-method}
\title{getSlot 
get the slot values}
\usage{
\S4method{getSlot}{stana}(x, slot)
}
\arguments{
\item{x}{stana object}

\item{slot}{slot name}
}
\description{
getSlot 
get the slot values
}
