% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stana.R
\name{getSNVID,stana-method}
\alias{getSNVID,stana-method}
\title{getSNVID
get SNV IDs from stana}
\usage{
\S4method{getSNVID}{stana}(x, candSp)
}
\arguments{
\item{x}{stana object}

\item{candSp}{species}
}
\description{
getSNVID
get SNV IDs from stana
}
