% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stana.R
\name{getTree,stana-method}
\alias{getTree,stana-method}
\title{getTree
get tree list from stana}
\usage{
\S4method{getTree}{stana}(x)
}
\arguments{
\item{x}{stana object}
}
\description{
getTree
get tree list from stana
}
