% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stana.R
\name{getTreePlot,stana-method}
\alias{getTreePlot,stana-method}
\title{getTreePlot
get tree plot list from stana}
\usage{
\S4method{getTreePlot}{stana}(x)
}
\arguments{
\item{x}{stana object}
}
\description{
getTreePlot
get tree plot list from stana
}
