% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTree.R
\name{inferAndPlotTree}
\alias{inferAndPlotTree}
\title{inferAndPlotTree}
\usage{
inferAndPlotTree(
  stana,
  species = NULL,
  cl = NULL,
  dist_method = "dist.ml",
  meta = NULL,
  layout = "circular",
  target = "fasta",
  IDs = NULL,
  use_point = FALSE,
  branch_col = "black",
  tree_args = list(),
  branch.length = "none",
  point_size = 2,
  deleteZeroDepth = TRUE
)
}
\arguments{
\item{stana}{stana object}

\item{species}{species to plot
If NULL, first species in fasta list is assigned}

\item{cl}{optional, cluster to plot}

\item{dist_method}{dist method in phangorn, default to dist.ml
if target is not fasta, ordinally `dist` method wil be used.}

\item{meta}{default to NULL, column name in `meta` slot}

\item{layout}{layout to be used in ggtree}

\item{target}{fasta, snp, or gene. default to fasta.}

\item{IDs}{if IDs is provided and target is snps or genes,
subset the matrix to this ID.}

\item{use_point}{use geom_point for plotting discrete metadata,
instead of geom_star.}

\item{tree_args}{passed to dist function in phangorn}

\item{branch.length}{branch length, default to "none", cladogram}

\item{point_size}{point size}

\item{deleteZeroDepth}{delete zero depth position}

\item{model}{dist.ml model}
}
\description{
infer the tree and plot.
If there is already a tree in the list from `treeList` slot, inferring will not be performed.
The function uses dist.ml and NJ for inference from FASTA format, and otherwise uses dist and NJ
for the inference. if metadata is available, use them to plot metadata around the tree.
}
