% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkProfile.R
\name{loadInStrain}
\alias{loadInStrain}
\title{loadInStrain}
\usage{
loadInStrain(
  compare_out_dir,
  sample_threshold,
  candidate_species,
  just_species = FALSE
)
}
\arguments{
\item{compare_out_dir}{output directory of compare.
need `output` in the directory}

\item{sample_threshold}{SNVs with at least this number of samples are included}

\item{candidate_species}{candidate species ID, e.g. `GUT_GENOME000024`}

\item{just_species}{if TRUE, returns the vector of species IDs}
}
\description{
For InStrain, you should provide `compare` output
with --bams options implmented from version 1.6.
It computes MAF of biallelic position.
Filters of popSNV or conSNV will be added.
}
