% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkProfile.R
\name{loadmetaSNV}
\alias{loadmetaSNV}
\title{loadmetaSNV}
\usage{
loadmetaSNV(metasnv_out_dir, cl = NULL)
}
\arguments{
\item{metasnv_out_dir}{output directory of merge_midas.py}

\item{cl}{named list of sample IDs}
}
\description{
Assess and store profile for species and return filtered species 
based on the number of samples for each category or whole population.
For metaSNV only.
}
