% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadProfile.R
\name{loadMIDAS}
\alias{loadMIDAS}
\title{loadMIDAS}
\usage{
loadMIDAS(
  midas_merge_dir,
  cl,
  filtType = "group",
  candSp = NULL,
  filtNum = 2,
  filtPer = 0.8,
  loadSummary = TRUE,
  loadDepth = TRUE,
  loadInfo = TRUE,
  geneType = "copynum"
)
}
\arguments{
\item{midas_merge_dir}{output directory of merge_midas.py}

\item{cl}{named list of sample IDs}

\item{filtType}{"whole" or "group"}

\item{candSp}{candidate species ID}

\item{filtNum}{The species with number above this threshold
for each category is returned}

\item{filtPer}{filter by fraction}

\item{loadSummary}{default to FALSE, load summary information.}

\item{loadDepth}{default to FALSE, load depth information.}

\item{loadInfo}{default to FALSE, load info information.}

\item{geneType}{"presabs" or "copynum"}
}
\description{
Assess and store profile for species and return filtered species 
based on the number of samples for each category or whole population.
For MIDAS only.
}
