% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkProfile.R
\name{loadMIDAS2}
\alias{loadMIDAS2}
\title{loadMIDAS2}
\usage{
loadMIDAS2(
  midas_merge_dir,
  cl,
  filtNum = 2,
  filtPer = 0.8,
  taxtbl = NULL,
  filtType = "group",
  geneType = "copynum"
)
}
\arguments{
\item{midas_merge_dir}{path to merged directory}

\item{cl}{named list of category for samples}

\item{filtNum}{the species with the samples above this number will be returned}

\item{filtPer}{filter by percentage}

\item{taxtbl}{tax table, row.names: 6-digits MIDAS2 ID and `GTDB species` column.}
}
\description{
load the MIDAS2 merge output.
The location to lz4 binary must be added to PATH.
Tax table can be loaded from downloaded MIDAS2 db directory (metadata.tsv).
If provided, additionally show tax names.
}
