% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF.R
\name{NMF}
\alias{NMF}
\title{NMF}
\usage{
NMF(
  stana,
  species,
  rank = 5,
  target = "KO",
  seed = 53,
  method = "snmf/r",
  deleteZeroDepth = TRUE,
  beta = 0.01,
  plotHeatmap = TRUE
)
}
\arguments{
\item{stana}{stana object}

\item{species}{candidate species ID}

\item{rank}{rank of NMF}

\item{target}{KO, gene or snv, default to KO}

\item{seed}{random seed}

\item{method}{NMF method, default to snmf/r}

\item{plotHeatmap}{plot the heatmap of presence of strain across sample}
}
\description{
decompose SNV, gene content, or gene family abundance to
strain x sample and strain x feature matrix.
}
