% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doGSEA.R
\name{obtainPositions}
\alias{obtainPositions}
\title{obtain_positions}
\usage{
obtainPositions(stana, candSp, geneID)
}
\arguments{
\item{stana}{stana object}

\item{candSp}{candidate species}

\item{geneID}{gene IDs}
}
\description{
obtain_positions
}
