% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF.R
\name{pathwayWithFactor}
\alias{pathwayWithFactor}
\title{pathwayWithFactor
convert KO matrix per factor obtained by NMF function
to pathway to factor matrix by summing the KOs in the pathway.}
\usage{
pathwayWithFactor(stana, species, tss = FALSE)
}
\arguments{
\item{stana}{stana boject}

\item{species}{species ID}

\item{tss}{perform total sum scaling to the resulting matrix}
}
\description{
pathwayWithFactor
convert KO matrix per factor obtained by NMF function
to pathway to factor matrix by summing the KOs in the pathway.
}
