% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCirclize.R
\name{plotCirclize}
\alias{plotCirclize}
\title{plotCirclize}
\usage{
plotCirclize(
  stana,
  candSp,
  genome_id,
  thresh_snp_gene = 5,
  type = "point",
  featList = list(),
  cols = c("steelblue", "tomato", "gold", "seagreen"),
  featThresh = 0,
  cex = 0.3,
  textCex = 0.5,
  bar_width = 10
)
}
\arguments{
\item{stana}{stana object}

\item{candSp}{candidate species ID}

\item{genome_id}{genome ID to plot}

\item{thresh_snp_gene}{include only genes with at least this number of SNVs}

\item{type}{one of `point`, `bar`, or `rect` each corresponding to circos functions}

\item{featList}{features to plot}

\item{cols}{color of each feature}

\item{featThresh}{for numeric features, the values above this value will be colored , default to zero.}

\item{cex}{point size}

\item{textCex}{text size}

\item{bar_width}{on circos.barplot}
}
\value{
draw circlize plot
}
\description{
Using `circlize` package, make a circular visualization of 
the relationship between genes and corresponding SNVs in genes.
}
