% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCoverage.R
\name{plotCoverage}
\alias{plotCoverage}
\title{plotCoverage}
\usage{
plotCoverage(stana, species, cl = NULL, pointSize = 5)
}
\arguments{
\item{stana}{stana object}

\item{species}{candidate species}

\item{cl}{named list of clusters}

\item{pointSize}{scatter point size}
}
\value{
ggplot2 object
}
\description{
plot the scatters of mean coverage across groups
for MIDAS1
}
