% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDist.R
\name{plotDist}
\alias{plotDist}
\title{plotDist}
\usage{
plotDist(
  stana,
  sp,
  cl = NULL,
  AAfunc = dist.ml,
  AAargs = list(),
  target = "snps",
  distMethod = "manhattan",
  distArg = list(),
  candidate = NULL
)
}
\arguments{
\item{stana}{stana object}

\item{sp}{species to be examined}

\item{cl}{named list of samples}

\item{AAfunc}{if choose `fasta`, provide function for calculating distance}

\item{AAargs}{provided to `AAfunc`}

\item{target}{tree, snps, genes, kos, fasta}

\item{distMethod}{distance method passed to dist() (default, manhattan)}

\item{distArg}{passed to `dist()}

\item{candidate}{list of IDs included for the calculation (like SNV IDs)}
}
\description{
Perform distance calculation based on the matrix
stored in stana, and plot the heatmap with grouping
information.
}
