% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doGSEA.R
\name{plotGSEA}
\alias{plotGSEA}
\title{plotGSEA}
\usage{
plotGSEA(
  stana,
  dataset_names = NULL,
  padjThreshold = 0.05,
  return_graph = FALSE,
  layout = "kk"
)
}
\arguments{
\item{stana}{stana object or list of stana object}

\item{dataset_names}{dataset names}

\item{padjThreshold}{threshold of adjusted p-values}

\item{return_graph}{return only the tbl_graph}

\item{layout}{graph layout in ggraph}
}
\value{
ggplot object
}
\description{
plot GSEA results
}
