% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotKEGGPathway.R
\name{plotKEGGPathway}
\alias{plotKEGGPathway}
\title{plotKEGGPathway}
\usage{
plotKEGGPathway(
  stana,
  species,
  pathway_id,
  cl = NULL,
  multi_scale = FALSE,
  eggnog = TRUE,
  kegg_name_match = "all",
  how = mean,
  eps = 0.01,
  color_list = NULL,
  only_ko = FALSE,
  summarize = FALSE,
  nudge_y = 0,
  point_mode = FALSE,
  sp_colors = NULL,
  point_size = 1,
  statMethod = "t"
)
}
\arguments{
\item{stana}{stana class object}

\item{species}{candidate species}

\item{pathway_id}{KEGG PATHWAY id}

\item{cl}{grouping (named list). If not specified, use stana@cl slot}

\item{multi_scale}{if TRUE, use ``ggh4x`` to plot multiple scales per species}

\item{eggnog}{use eggNOG annotation in stana object}

\item{kegg_name_match}{passed to node_numeric function}

\item{how}{how to concatenate abundance per KO}

\item{eps}{pseudovalue added when claculating log2 fold change}

\item{color_list}{color list if multi_scale=TRUE,
passed to scale_fill_multi()}

\item{only_ko}{only calculates KO and return stana}

\item{summarize}{summarize the value if multiple species are set ("+")}

\item{nudge_y}{nudge the y positions in point_mode}

\item{point_mode}{default to FALSE, show whether the corresponding KOs
are present in species by points}

\item{sp_colors}{species colors used in point_mode}

\item{point_size}{point size in point_mode}

\item{statMethod}{how to colorize the nodes}
}
\value{
list of plots or plot
}
\description{
plot the KEGG pathway
The function uses `cl` slot and output statistics based on the groups.
If the groups number is above two, the sum value across the group 
will be used (not using grouping information).
}
