% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMAF.R
\name{plotMAF}
\alias{plotMAF}
\title{plotMAF}
\usage{
plotMAF(stana, species, SNV, cl = NULL, deleteZeroDepth = TRUE)
}
\arguments{
\item{stana}{stana object}

\item{species}{candidate species}

\item{SNV}{snv position name}

\item{cl}{named list of clusters}

\item{deleteZeroDepth}{delete position with zero depth}
}
\value{
ggplot
}
\description{
plot the box plot of MAF for specificed SNV position
}
