% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMAF.R
\name{plotMAFHist}
\alias{plotMAFHist}
\title{plotMAFHist}
\usage{
plotMAFHist(stana, species)
}
\arguments{
\item{stana}{stana object}

\item{species}{candidate species}
}
\value{
ggplot
}
\description{
plot the distribution of MAF for specificed species
}
