% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPCA.R
\name{plotPCA}
\alias{plotPCA}
\title{plotPCA}
\usage{
plotPCA(
  stana,
  species,
  cl,
  target = "snps",
  pointSize = 5,
  cols = palette(),
  useBlend = FALSE
)
}
\arguments{
\item{species}{candidate species}

\item{cl}{named list of clusters}

\item{pointSize}{scatter point size}

\item{staba}{output of midas merge}
}
\description{
plot the PCA results from snp or gene table
}
