% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plotSNVSummary}
\alias{plotSNVSummary}
\title{plotSNVSummary}
\usage{
plotSNVSummary(stana, sp, param = "mean_coverage", perSample = FALSE)
}
\arguments{
\item{stana}{stana object}

\item{sp}{species_id}

\item{param}{parameter to plot}

\item{perSample}{plot using no sample information}
}
\value{
ggplot object
}
\description{
plotSNVSummary
}
