% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTree.R
\name{plotTree}
\alias{plotTree}
\title{plotTree}
\usage{
plotTree(
  stana,
  species = NULL,
  cl = NULL,
  dist_method = "dist.ml",
  tree_args = list(),
  branch.length = "none"
)
}
\arguments{
\item{stana}{stana object}

\item{species}{species to plot
If NULL, first species in fasta list is assigned}

\item{cl}{optional, cluster to plot}

\item{dist_method}{dist method in phangorn, default to dist.ml}

\item{tree_args}{passed to dist function in phangorn}

\item{branch.length}{branch length, default to "none", cladogram}

\item{model}{dist.ml model}
}
\description{
infer the tree and plot
use dist.ml and NJ for inference
}
