% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankComponents.R
\name{rankComponents}
\alias{rankComponents}
\title{rankComponents}
\usage{
rankComponents(
  stana,
  pid,
  candSp = NULL,
  cl = NULL,
  eps = 0.01,
  how = sum,
  zeroPerc = 0,
  rankMethod = "modt",
  statHow = mean
)
}
\arguments{
\item{stana}{stana object}

\item{pid}{KEGG pathway ID}

\item{candSp}{candidate species}

\item{cl}{if NULL, grouping in stana is used}

\item{eps}{pseudovalue added if log is taken}

\item{how}{how to combine multiple variables, default to sum}

\item{rankMethod}{how to rank genes}

\item{statHow}{how to aggregate the multiple statistical values}
}
\description{
Returns the rank of the components in the graph
based on the statistical values inside the graph.
}
