% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{returnGenes}
\alias{returnGenes}
\title{returnGenes
return corresponding genes queried by  SNV position}
\usage{
returnGenes(stana, species, snvs)
}
\arguments{
\item{stana}{stana object}

\item{species}{species name}

\item{snvs}{snv name}
}
\value{
named vector
}
\description{
returnGenes
return corresponding genes queried by  SNV position
}
