% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doGSEA.R
\name{reverseAnnot}
\alias{reverseAnnot}
\title{reverse_annot}
\usage{
reverseAnnot(stana, candSp, candidate, col = "KEGG_ko")
}
\arguments{
\item{stana}{stana object}

\item{candSp}{candidate species}

\item{candidate}{candidate gene family ids}

\item{col}{column to use in eggNOG mapper annotation}
}
\description{
reverse_annot
}
