% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scaler.NNLM}
\alias{scaler.NNLM}
\title{scaler.NNLM}
\usage{
scaler.NNLM(nmf, target = "coef")
}
\arguments{
\item{nmf}{NNLM object}

\item{target}{coef or basis}
}
\value{
list
}
\description{
scaler.NNLM
}
