% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stana.R
\name{setAnnotation}
\alias{setAnnotation}
\title{setAnnotation}
\usage{
setAnnotation(stana, annotList)
}
\arguments{
\item{stana}{stana object}

\item{annotList}{named list of the path to eggNOG annotation file
(*emapper.annotations)}
}
\value{
stana object
}
\description{
setAnnotation
}
