% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stana.R
\name{setMap}
\alias{setMap}
\title{setMap}
\usage{
setMap(stana, candSp, map)
}
\arguments{
\item{stana}{stana object}

\item{candSp}{species ID}

\item{map}{mapping data.frame, two column consisting of ID and value.
The function internally changes the column name.
ID is gene ID in `genes` slot, and value indicates orthology ID (like K00001)}
}
\value{
stana object
}
\description{
setMap
}
