% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setMetadata.R
\name{setMetadata}
\alias{setMetadata}
\title{setMetadata}
\usage{
setMetadata(stana, meta)
}
\arguments{
\item{stana}{stana}

\item{meta}{metadata (data.frame)
row.name as ID of the trees}
}
\value{
stana object
}
\description{
setMetadata
}
