% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stana.R
\name{setSlot,stana-method}
\alias{setSlot,stana-method}
\title{setSlot
general method for setting the slot}
\usage{
\S4method{setSlot}{stana}(x, y, z)
}
\arguments{
\item{x}{stana object}

\item{y}{slot name}

\item{z}{set variable}
}
\description{
setSlot
general method for setting the slot
}
