% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stana.R
\name{setSlot}
\alias{setSlot}
\title{setSlot
general method for setting the slot}
\usage{
setSlot(x, y, z)
}
\arguments{
\item{x}{stana object}

\item{y}{slot name}

\item{z}{set variable}
}
\description{
setSlot
general method for setting the slot
}
