% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stana.R
\name{setSNVID,stana-method}
\alias{setSNVID,stana-method}
\title{setSNVID}
\usage{
\S4method{setSNVID}{stana}(x, sp, IDs)
}
\arguments{
\item{x}{stana object}

\item{sp}{species ID}

\item{IDs}{SNV ID}
}
\description{
setSNVID
}
