% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setTree}
\alias{setTree}
\title{setTree}
\usage{
setTree(stana, species, tre)
}
\arguments{
\item{stana}{stana object}

\item{species}{species ID}

\item{tre}{tree to be set}
}
\value{
stana
}
\description{
setTree
}
