% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stana.R
\name{siteFilter,stana-method}
\alias{siteFilter,stana-method}
\title{siteFilter}
\usage{
\S4method{siteFilter}{stana}(x, sp, exp)
}
\arguments{
\item{x}{stana object}

\item{sp}{species ID}

\item{exp}{expression for filtering the snps info}
}
\description{
siteFilter
}
