% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stana.R
\name{siteFilter}
\alias{siteFilter}
\title{siteFilter}
\usage{
siteFilter(x, sp, exp)
}
\arguments{
\item{x}{stana object}

\item{sp}{species ID}

\item{exp}{expression for filtering the snps info}
}
\description{
siteFilter
}
