% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stana.R
\docType{class}
\name{stana-class}
\alias{stana-class}
\title{Note that gene summary and depth is not loaded in MIDAS and MIDAS2.}
\description{
Note that gene summary and depth is not loaded in MIDAS and MIDAS2.
}
\section{Slots}{

\describe{
\item{\code{snpsInfo}}{snp info}

\item{\code{snpsDepth}}{snp depth}

\item{\code{snpsSummary}}{snp summary}

\item{\code{kos}}{stores ko table}

\item{\code{type}}{type of pipeline, such as `MIDAS`}

\item{\code{eggNOG}}{list of path for eggNOG mapper v2 results for species}

\item{\code{map}}{slot storing the mapping data.frame for gene ID and orthology}
}}

