% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InStrainFuncs.R
\name{strainClusterHeatmap}
\alias{strainClusterHeatmap}
\title{strainClusterHeatmap}
\usage{
strainClusterHeatmap(stana, species, cl = NULL, heatmapArgs = list())
}
\arguments{
\item{stana}{stana object}

\item{species}{interesting species}

\item{cl}{grouping, if NULL, it automatically obtain grouping from stana object}

\item{heatmapArgs}{list of parameters passed to Heatmap()}
}
\value{
Heatmap by ComplexHeatmap
}
\description{
Draw a heatmap representing presence/absence of strain cluster
}
