% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stana.R
\name{summary,stana-method}
\alias{summary,stana-method}
\title{summary
print summary information}
\usage{
\S4method{summary}{stana}(object, ...)
}
\arguments{
\item{object}{stana object}

\item{...}{other arguments}
}
\description{
summary
print summary information
}
