% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stana.R
\name{summary}
\alias{summary}
\title{summary
print summary information}
\usage{
summary(object, ...)
}
\arguments{
\item{object}{stana object}

\item{...}{other arguments}
}
\description{
summary
print summary information
}
