% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doGSEA.R
\name{addGeneAbundance}
\alias{addGeneAbundance}
\title{addGeneAbundance}
\usage{
addGeneAbundance(
  stana,
  candSp,
  IDs,
  target = "KO",
  how = sum,
  newCol = "gene",
  disc = NULL,
  discNumeric = TRUE
)
}
\arguments{
\item{disc}{discretize the abundance by the threshold. function for calculating
threshold, like {median}}
}
\description{
Add the specified gene copy number to metadata
}
