% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkEGGNOG.R
\name{checkEGGNOG}
\alias{checkEGGNOG}
\title{checkEGGNOG}
\usage{
checkEGGNOG(annot_file, ret = "all", checkIDs = NULL, fill = TRUE)
}
\arguments{
\item{annot_file}{path to eggnog-mapper annotation}

\item{ret}{default to all}

\item{checkIDs}{only return functions related to IDs}
}
\description{
When the interesting genes (e.g. differentially abundant between category) are found,
one can query eggNOG-mapper for its functionality.
This function assumes one performs eggNOG-mapper to centroids.ffn of candidate species, and
parse the annotation results.
}
\details{
Choices of IDs in `ret`:
"COG_category" "Description" "Preferred_name" "GOs"           
"EC"             "KEGG_ko"        "KEGG_Pathway"   "KEGG_Module"    "KEGG_Reaction" 
"KEGG_rclass"    "BRITE"          "KEGG_TC"        "CAZy"           "BiGG_Reaction" 
"PFAMs"
}
