% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineSeqs.R
\name{combineSeqs}
\alias{combineSeqs}
\title{combineSeqs}
\usage{
combineSeqs(stana_list, species, argList = list(), output_seq = FALSE)
}
\arguments{
\item{stana_list}{stana list}

\item{species}{species ID}

\item{argList}{args to passed to consensusSeq}

\item{output_seq}{output just the sequence}
}
\value{
new stana object
}
\description{
(experimental) use common SNV across multiple datasets to
infer the MSA, the same species and same database must be used.
}
