% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensusSeq.R, R/consensusSeqFast.R
\name{consensusSeq}
\alias{consensusSeq}
\alias{consensusSeqGeneral}
\alias{consensusSeqMIDAS2}
\alias{consensusSeqMIDAS1}
\title{consensusSeq}
\usage{
consensusSeq(stana, species = NULL, argList = list())

consensusSeqGeneral(
  stana,
  species = NULL,
  cl = NULL,
  keep_samples = NULL,
  output_seq = FALSE,
  tree = FALSE,
  return_mat = FALSE,
  allele_columns = c("major_allele", "minor_allele"),
  verbose = FALSE
)

consensusSeqMIDAS2(
  stana,
  species = NULL,
  output_seq = FALSE,
  mean_depth = 0,
  fract_cov = 0,
  site_depth = 5,
  site_ratio = 5,
  site_maf = 0.01,
  allele_support = 0.5,
  site_prev = 0.9,
  locus_type = "CDS",
  site_list = NULL,
  cl = NULL,
  max_sites = Inf,
  tree = FALSE,
  keep_samples = NULL,
  return_mat = FALSE,
  verbose = FALSE
)

consensusSeqMIDAS1(
  stana,
  species = NULL,
  mean_depth = 0,
  fract_cov = 0,
  site_depth = 5,
  site_ratio = 5,
  site_maf = 0.01,
  allele_support = 0.5,
  site_prev = 0.9,
  cl = NULL,
  max_sites = Inf,
  tree = FALSE,
  locus_type = "CDS",
  keep_samples = NULL,
  verbose = FALSE,
  site_list = NULL,
  return_mat = FALSE,
  output_seq = FALSE
)
}
\arguments{
\item{stana}{stana object}

\item{species}{candidate species vector}

\item{argList}{parameters, passed to corresponding functions}

\item{cl}{cluster, if plot cladogram}

\item{keep_samples}{samples to kept in inference}

\item{output_seq}{whether to output actual FASTA file}

\item{tree}{whether or not to perform tree inference using dist.ml()
and NJ() in default parameters}

\item{return_mat}{return matrix of characters}

\item{allele_columns}{columns specifying major and minor allele}

\item{verbose}{output current status}

\item{mean_depth}{parameter for sample filtering
mean coverage across the sample}

\item{fract_cov}{parameter for sample filtering
fraction of coverage per sample}

\item{site_depth}{parameter for site filtering
minimum site depth}

\item{site_ratio}{parameter for site filtering
depth divided by mean coverage}

\item{site_maf}{parameter for site filtering
the minimum maf to be included}

\item{allele_support}{parameter for site filtering}

\item{site_prev}{parameter for site filtering
site prevalence across samples}

\item{locus_type}{locus type to be included, default to CDS}

\item{site_list}{site list to be included}

\item{max_sites}{default to Inf, max sites to be retained}
}
\description{
Output consensus sequences from merged SNV output.
Optionally, return phylogenetic tree inferred by `phangorn`.
If specified cluster of samples, additionally returns plot by `ggtree`.

Perform consensus sequence calling using MAF table.
The function does not filter the sites or samples, so the prefiltering based on 
depth or samples profiled is needed. The function assumes non-profiled site value coded as -1.
}
