% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensusSeq.R
\name{consensusSeq}
\alias{consensusSeq}
\title{consensusSeq}
\usage{
consensusSeq(stana, species = NULL, argList = list())
}
\arguments{
\item{stana}{stana object}

\item{species}{species vectors}

\item{argList}{parameters, passed to corresponding functions}
}
\description{
Output consensus sequences from merged SNV output.
Optionally, return phylogenetic tree inferred by `phangorn`.
If specified cluster of samples, additionally returns plot by `ggtree`.
}
