% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensusSeqFast.R
\name{consensusSeqGeneral}
\alias{consensusSeqGeneral}
\title{consensusSeqGeneral}
\usage{
consensusSeqGeneral(
  stana,
  species = NULL,
  cl = NULL,
  keep_samples = NULL,
  output_seq = FALSE,
  tree = FALSE,
  return_mat = FALSE,
  allele_columns = c("major_allele", "minor_allele"),
  verbose = FALSE
)
}
\arguments{
\item{stana}{stana obj}

\item{species}{candidate species vector}

\item{cl}{cluster, if plot cladogram}

\item{keep_samples}{the samples to keep}

\item{output_seq}{whether to output actual FASTA file}

\item{tree}{if TRUE, perform tree inference}

\item{return_mat}{return matrix of characters}

\item{allele_columns}{columns specifying major and minor allele}

\item{verbose}{output current status}
}
\description{
consensusSeqGeneral
}
