% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensusSeqFast.R
\name{consensusSeqGeneral}
\alias{consensusSeqGeneral}
\title{consensusSeqGeneral}
\usage{
consensusSeqGeneral(
  stana,
  species = NULL,
  output_seq = FALSE,
  tree = FALSE,
  return_mat = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{stana}{stana obj}

\item{species}{candidate species vector}

\item{output_seq}{whether to output actual FASTA file}

\item{tree}{if perform tree inference}

\item{return_mat}{return matrix of characters}

\item{verbose}{output current status}

\item{mean_depth}{parameter for filtering}

\item{fract_cov}{parameter for filtering}

\item{site_depth}{parameter for filtering}

\item{site_ratio}{parameter for filtering}

\item{site_maf}{parameter for filtering}

\item{allele_support}{parameter for filtering}

\item{site_prev}{parameter for filtering}

\item{cl}{cluster, if plot cladogram}

\item{max_sites}{default to Inf}

\item{keep_samples}{currently not implemented}

\item{exclude_samples}{currently not implemented}

\item{rand_samples}{currently not implemented}

\item{max_samples}{currently not implemented}

\item{locus_type}{locus type to be included, default to CDS}

\item{site_list}{site list to be included}
}
\description{
consensusSeqGeneral
}
