% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doAdonis.R
\name{doAdonis}
\alias{doAdonis}
\title{doAdonis}
\usage{
doAdonis(
  stana,
  specs,
  cl,
  target = "snps",
  formula = NULL,
  distMethod = "manhattan",
  maj = FALSE,
  argList = list()
)
}
\arguments{
\item{stana}{stana object}

\item{specs}{species to be examined}

\item{cl}{named list of samples}

\item{target}{snps, presabs, or copynum}

\item{formula}{pass to adonis2, specify distance matrix as d.}

\item{distMethod}{distance method passed to dist() (default, manhattan)}

\item{maj}{major allele distance}

\item{argList}{parameters passed to adonis2}
}
\description{
Perform PERMANOVA on distance matrix based 
on SNV frequency or gene matrix.
Named list of samples are to be provided.
Note that this function currently performs 
comparison using one category.
for complex design, one should do it manually.
}
