% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doBoruta.R
\name{doBoruta}
\alias{doBoruta}
\title{doBoruta}
\usage{
doBoruta(stana, cand, cl, target = "snps", mat = NULL)
}
\arguments{
\item{stana}{stana object}

\item{cand}{candidate species}

\item{cl}{named list of category}

\item{target}{default to snps}

\item{mat}{if target is not snps, provide preprocessed gene matrix
otherwise the raw gene matrix is used.}
}
\description{
feature selection of classification between category
using snv frequency and gene matrix
}
