% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doGSEA.R
\name{doGSEA}
\alias{doGSEA}
\title{doGSEA}
\usage{
doGSEA(stana, candSp = NULL, cl = NULL, eps = 0.01, how = mean, zeroPerc = 0)
}
\value{
GSEA results from clusterProfiler
}
\description{
By default this uses t-statistics for ranking of the genes.
}
