% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doGSEA.R
\name{doGSEA}
\alias{doGSEA}
\title{doGSEA}
\usage{
doGSEA(
  stana,
  candSp = NULL,
  cl = NULL,
  eps = 0.01,
  how = mean,
  zeroPerc = 0,
  rankMethod = "modt",
  target = "pathway"
)
}
\arguments{
\item{zeroPerc}{genes >= the percentage of count zero sample will be excluded.
Default to zero, not recommended in GSEA}
}
\value{
GSEA results from clusterProfiler
}
\description{
Based on KEGG database, GSEA will be performed by the function.
clusterProfiler::GSEA is used.
By default this uses moderated t-statistics for ranking of the genes.
}
