% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkEGGNOG.R
\name{drawEGGNOG}
\alias{drawEGGNOG}
\title{drawEGGNOG}
\usage{
drawEGGNOG(annot_file, geneIDs, candPlot)
}
\arguments{
\item{annot_file}{annotation file}

\item{geneIDs}{geneIDs}

\item{candPlot}{vector consisting of 
"cog","ec","ko","kegg","module","reaction"}
}
\description{
draw the graph representing the eggNOG annotation
}
