% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportInteractive.R
\name{exportInteractive}
\alias{exportInteractive}
\title{exportInteractive}
\usage{
exportInteractive(
  stana,
  out = ".",
  db = "uhgg",
  calcko = TRUE,
  calctree = TRUE,
  species = NULL
)
}
\arguments{
\item{stana}{stana object}

\item{out}{output directory
export the stana results for the inspection in interactive
application}

\item{db}{db used to profile 'uhgg' or 'gtdb'}

\item{calcko}{calculate KO abundance}

\item{calctree}{calculate consensus tree}

\item{species}{calk}
}
\value{
output the results to specified directory
}
\description{
exportInteractive
}
