% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportInteractive.R
\name{exportInteractive}
\alias{exportInteractive}
\title{exportInteractive}
\usage{
exportInteractive(
  stana,
  out = ".",
  db = "uhgg",
  calcko = TRUE,
  calctree = TRUE,
  dataset_name = NULL,
  species = NULL
)
}
\arguments{
\item{stana}{stana object of type MIDAS2}

\item{out}{output directory}

\item{db}{db used to profile 'uhgg' or 'gtdb'}

\item{calcko}{calculate KO abundance}

\item{calctree}{calculate consensus tree}

\item{dataset_name}{dataset name}

\item{species}{candidate species, default to all the species}
}
\value{
output the results to specified directory
}
\description{
export the current stana object to interactive application 
for the convenient analysis and visualization
if `cl` and `meta` slot is filled, `meta` slot is exported.
}
