% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferAndPlotTree.R
\name{inferAndPlotTree}
\alias{inferAndPlotTree}
\title{inferAndPlotTree}
\usage{
inferAndPlotTree(
  stana,
  species = NULL,
  cl = NULL,
  dist_method = "dist.ml",
  meta = NULL,
  layout = "circular",
  target = "fasta",
  IDs = NULL,
  use_point = FALSE,
  branch_col = "black",
  tree_args = list(),
  branch.length = "none",
  point_size = 2,
  subset_samples = NULL,
  deleteZeroDepth = FALSE,
  treeFun = "upgma",
  tree_only = FALSE
)
}
\arguments{
\item{stana}{stana object}

\item{species}{species to plot
If NULL, first species in fasta list is assigned}

\item{cl}{optional, cluster to plot}

\item{dist_method}{dist method in phangorn, default to dist.ml
if target is not fasta, ordinally `dist` method wil be used.}

\item{meta}{default to NULL, column name in `meta` slot}

\item{layout}{layout to be used in ggtree}

\item{target}{fasta, snp, or gene. default to fasta.}

\item{IDs}{if IDs is provided and target is snps or genes,
subset the matrix to this ID.}

\item{use_point}{use geom_point for plotting discrete metadata,
instead of geom_star.}

\item{branch_col}{branch color, if this option is in the column name of metadata,
use this variable as coloring. Default to black.}

\item{tree_args}{passed to dist function}

\item{branch.length}{branch length, default to "none", cladogram}

\item{point_size}{point size}

\item{subset_samples}{subset samples (for matrix only)}

\item{deleteZeroDepth}{delete zero depth position}

\item{treeFun}{tree inferring function in phangorn
if `FastTree` is specified, FastTree will be called.
In this case, FastTree should be in PATH.}

\item{tree_only}{return tree only instead of stana object}
}
\value{
stana object
}
\description{
infer the tree and plot.
If there is already a tree in the list from `treeList` slot, inferring will not be performed.
The function uses dist.ml and NJ for inference from FASTA format, and otherwise uses dist and NJ
for the inference. if metadata is available, use them to plot metadata around the tree.
}
