% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadProfile.R
\name{loadInStrain}
\alias{loadInStrain}
\title{loadInStrain}
\usage{
loadInStrain(
  compare_out_dir,
  candidate_species,
  cl = NULL,
  just_species = FALSE,
  fill_na = TRUE,
  skip_pool = TRUE
)
}
\arguments{
\item{compare_out_dir}{output directory of compare.
need `output` in the directory}

\item{candidate_species}{candidate species ID, e.g. `GUT_GENOME000024`}

\item{cl}{named list of grouping}

\item{just_species}{if TRUE, returns the vector of species IDs
(needs pooled SNV data, otherwise use genome wide compraison table)}

\item{fill_na}{fill NA in resulting data.frame by -1}

\item{skip_pool}{skip pooled results loading}
}
\description{
For InStrain, you should provide `compare` output
It can import outputs with --bams options implmented from version 1.6.
It computes MAF of var_base per position and return the frequency matrix.
Note it considers con_base and var_base in info table for calculation.
}
