% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadMIDAS2Refine.R
\name{loadMIDAS2}
\alias{loadMIDAS2}
\title{loadMIDAS2}
\usage{
loadMIDAS2(
  midas_merge_dir,
  cl = NULL,
  filtNum = 2,
  db = "gtdb",
  only_stat = FALSE,
  filtBy = "snps",
  filtPer = 0.8,
  candSp = NULL,
  filtType = "group",
  geneType = "copynum",
  loadSummary = TRUE,
  loadInfo = TRUE,
  loadDepth = TRUE
)
}
\arguments{
\item{midas_merge_dir}{path to merged directory}

\item{cl}{named list of category for samples}

\item{filtNum}{the species with the samples above this number will be returned}

\item{db}{data base that was used to profile, default to gtdb.}

\item{only_stat}{only samples per species is returned (snpStat and geneStat)}

\item{filtBy}{filter by snps number or genes number (default to snps)}

\item{filtPer}{filter by fraction}

\item{candSp}{candidate species ID}

\item{filtType}{"whole" or "group"}

\item{geneType}{"presabs" or "copynum"}

\item{loadSummary}{default to FALSE, load summary information.}

\item{loadInfo}{default to FALSE, load info information.}

\item{loadDepth}{default to FALSE, load depth information.}
}
\value{
stana object
}
\description{
load the MIDAS2 merge command output.
The location to lz4 binary must be added to PATH.
Taxonomy table can be loaded from downloaded MIDAS2 db directory (metadata.tsv).
If provided, additionally show tax names.
}
