% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadProfile.R
\name{loadMIDAS2}
\alias{loadMIDAS2}
\title{loadMIDAS2}
\usage{
loadMIDAS2(
  midas_merge_dir,
  cl,
  filtNum = 2,
  db = "gtdb",
  filtPer = 0.8,
  taxtbl = NULL,
  candSp = NULL,
  filtType = "group",
  geneType = "copynum",
  loadSummary = FALSE,
  loadInfo = FALSE,
  loadDepth = FALSE
)
}
\arguments{
\item{midas_merge_dir}{path to merged directory}

\item{cl}{named list of category for samples}

\item{filtNum}{the species with the samples above this number will be returned}

\item{db}{data base that was used to profile, default to gtdb.}

\item{filtPer}{filter by fraction}

\item{taxtbl}{tax table, row.names: 6-digits MIDAS2 ID and `GTDB species` (gtdb)
or `Lineage` column.}

\item{candSp}{candidate species ID}

\item{filtType}{"whole" or "group"}

\item{geneType}{"presabs" or "copynum"}

\item{loadSummary}{default to FALSE, load summary information.}

\item{loadInfo}{default to FALSE, load info information.}

\item{loadDepth}{default to FALSE, load depth information.}
}
\description{
load the MIDAS2 merge command output.
The location to lz4 binary must be added to PATH.
Taxonomy table can be loaded from downloaded MIDAS2 db directory (metadata.tsv).
If provided, additionally show tax names.
}
