% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF.R
\name{NMF}
\alias{NMF}
\title{NMF}
\usage{
NMF(
  stana,
  species,
  rank = 3,
  target = "KO",
  seed = 53,
  method = "snmf/r",
  deleteZeroDepth = FALSE,
  beta = 0.01,
  estimate = FALSE,
  estimate_range = 1:6,
  nnlm_flag = FALSE,
  nnlm_args = list(),
  nnlm_na_perc = 0.3
)
}
\arguments{
\item{stana}{stana object}

\item{species}{candidate species ID}

\item{rank}{rank of NMF}

\item{target}{KO, gene or snv, default to KO}

\item{seed}{random seed}

\item{method}{NMF method, default to snmf/r}

\item{deleteZeroDepth}{when snv matrix is used, this option filters the
positions with zero-depth (indicated by -1)}

\item{beta}{argument to be passed to NMF function}

\item{estimate}{estimate rank}

\item{estimate_range}{range of ranks for the estimation}

\item{nnlm_flag}{if TRUE, use NNLM package which can handle missing values in the data.
The different approach for estimating k is taken when estimate is TRUE.
When estimate is TRUE, only the error matrix is returned.}

\item{nnlm_args}{arguments passed to NNLM functions}

\item{nnlm_na_perc}{NA frequency when the cross-validation is performed.}
}
\description{
decompose SNV, gene content, or gene family abundance to
factor x sample and factor x feature matrix.
}
