% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF.R
\name{pathwayWithFactor}
\alias{pathwayWithFactor}
\title{pathwayWithFactor
convert KO matrix per factor obtained by NMF function
to pathway to factor matrix by summing the KOs in the pathway.}
\usage{
pathwayWithFactor(
  stana,
  species,
  tss = FALSE,
  change_name = FALSE,
  summarize = sum,
  mat = NULL
)
}
\arguments{
\item{stana}{stana boject}

\item{species}{species ID}

\item{tss}{perform total sum scaling to the resulting matrix}

\item{change_name}{change pathway names to description}

\item{summarize}{summarizing function, default to base::sum}

\item{mat}{other matrix than the basis of NMF}
}
\description{
pathwayWithFactor
convert KO matrix per factor obtained by NMF function
to pathway to factor matrix by summing the KOs in the pathway.
}
