% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF.R
\name{plotAbundanceWithinSpecies}
\alias{plotAbundanceWithinSpecies}
\title{plotAbundanceWithinSpecies}
\usage{
plotAbundanceWithinSpecies(stana, species, tss = TRUE, return_data = FALSE)
}
\arguments{
\item{stana}{stana object}
}
\description{
plotAbundanceWithinSpecies
}
