% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCirclize.R
\name{plotCirclize}
\alias{plotCirclize}
\title{plotCirclize}
\usage{
plotCirclize(
  stana,
  candSp,
  genomeId,
  include_gene = NULL,
  thresh_snp_gene = 5,
  featList = list(),
  cols = c("tomato", "steelblue", "gold", "seagreen"),
  controlColor = "steelblue",
  MAF = FALSE,
  cl = NULL,
  featCircos = list(),
  showGeneName = TRUE,
  returnRawDf = FALSE,
  only_genome_id = FALSE,
  featThresh = 0,
  cex = 0.3,
  textCex = 0.5,
  bar_width = 10,
  contPalette = c("steelblue", "tomato"),
  discPalette = "Dark2",
  convertGeneID = FALSE
)
}
\arguments{
\item{stana}{stana object}

\item{candSp}{candidate species ID}

\item{genomeId}{genome ID to plot}

\item{include_gene}{include these gene IDs, ignoring the thesh_snp_gene argument}

\item{thresh_snp_gene}{include only genes with at least this number of SNVs}

\item{featList}{features to plot}

\item{cols}{color of each feature}

\item{controlColor}{control (not reaching `featThresh`) color in points and barplot}

\item{MAF}{plot MAF for each group}

\item{cl}{cluster information (named list)}

\item{featCircos}{which type of circos to use in the corresponding feature}

\item{showGeneName}{show gene ID on the plot}

\item{returnRawDf}{return the raw data frame used in the plot}

\item{only_genome_id}{show only the genome IDs}

\item{featThresh}{for numeric features, the values above this value will be colored , default to zero.}

\item{cex}{point size}

\item{textCex}{text size}

\item{bar_width}{on circos.barplot}

\item{contPalette}{palette for continuous scale}

\item{discPalette}{palette for discrete scale}

\item{convertGeneID}{convert the gene id to KO id in eggNOG annotation}
}
\value{
draw circlize plot
}
\description{
Using `circlize` package, make a circular visualization of 
the relationship between genes and corresponding SNVs in genes.
}
