% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCoverage.R
\name{plotCoverage}
\alias{plotCoverage}
\title{plotCoverage}
\usage{
plotCoverage(midas_merge_dir, species, cl, pointSize = 5)
}
\arguments{
\item{midas_merge_dir}{output of midas merge}

\item{species}{candidate species}

\item{cl}{named list of clusters}

\item{pointSize}{scatter point size}
}
\description{
plot the scatters of mean coverage across groups
for MIDAS1
}
