% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGenes.R
\name{plotGenes}
\alias{plotGenes}
\title{plotGenes}
\usage{
plotGenes(
  stana,
  species,
  geneID,
  target = "genes",
  cl = NULL,
  return_df = FALSE,
  geom = geom_violin
)
}
\arguments{
\item{stana}{stana object}

\item{species}{candidate species}

\item{geneID}{gene ID to be plotted}

\item{target}{genes or KOs (gene families)}

\item{cl}{named list of clusters}

\item{return_df}{return data only}

\item{geom}{default to geom_violin, can be changed to the geom like `geom_boxplot`}
}
\value{
ggplot
}
\description{
Plot the plot for specific genes
}
\details{
Plot the gene copy number between groups based on the specified geom.
The default is to use gene copy number, but gene family abundance can also be specified.
}
