% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGenes.R
\name{plotGenes}
\alias{plotGenes}
\title{plotGenes}
\usage{
plotGenes(
  stana,
  species,
  geneID,
  target = "genes",
  cl = NULL,
  return_df = FALSE
)
}
\arguments{
\item{stana}{stana object}

\item{species}{candidate species}

\item{geneID}{gene ID to be plotted}

\item{target}{genes or KOs (gene families)}

\item{cl}{named list of clusters}

\item{return_df}{return data only}
}
\value{
ggplot
}
\description{
plot the violin plot for specific genes
}
