% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHeatmap.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{plotHeatmap}
\usage{
plotHeatmap(
  stana,
  sp,
  cl = NULL,
  km = 10,
  mat = NULL,
  seed = 1,
  fnc = "KEGG_Pathway"
)
}
\arguments{
\item{stana}{stana object}

\item{sp}{candidate species}

\item{cl}{grouping named list}

\item{km}{k-means param for splitting gene}

\item{mat}{if customized matrix is to be used. Row will be gene names
and column sample names}

\item{fnc}{One of `KEGG_Pathway` or `KEGG_Module`, when eggNOG annotation is used}
}
\description{
plot a heatmap with functional annotation by simplifyEnrichment
Typically, MIDAS and MIDAS2 output are used
}
