% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHeatmap.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{plotHeatmap}
\usage{
plotHeatmap(
  stana,
  sp,
  cl = NULL,
  k = 10,
  mat = NULL,
  geneID = NULL,
  variable = 0,
  fnc = "KEGG_Pathway",
  removeHigh = TRUE,
  removeAdditional = NULL,
  max_words = 10,
  filter_zero_frac = 0.8,
  filter_max_frac = 0,
  filter_max_value = 50
)
}
\arguments{
\item{stana}{stana object}

\item{sp}{candidate species}

\item{cl}{grouping named list}

\item{k}{k-means param for splitting gene}

\item{mat}{if customized matrix is to be used. Row will be gene names
and column sample names}

\item{geneID}{gene id to subset}

\item{variable}{If specified other than 0, subset to top-{variable} variation gene numbers.}

\item{fnc}{One of `KEGG_Pathway` or `KEGG_Module`, when eggNOG annotation is used}

\item{removeHigh}{remove high frequent words (preset)}

\item{removeAdditional}{remove additional words specified}

\item{max_words}{max words to plot}

\item{filter_zero_frac}{genes with zero abundance over fraction of samples as this value
are removed before sample filtering. As typically gene matrix is large, for further filtering, please use `mat` option}

\item{filter_max_frac}{remove genes with values below `filter_max_value` in this fraction of sample}

\item{filter_max_value}{max value for copy numbers (default to 50), coupled with filter_max_frac}
}
\description{
Plot a heatmap of gene copy number.
}
\details{
Plot a heatmap with functional annotation by simplifyEnrichment.
The annotations are provided by `eggNOG` slot or `map` slot. If loaded type is MIDAS,
The function automatically fetches the functional annotation from PATRIC server.
}
