% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotKEGGPathway.R
\name{plotKEGGPathway}
\alias{plotKEGGPathway}
\title{plotKEGGPathway}
\usage{
plotKEGGPathway(
  stana,
  species,
  pathway_id,
  cl = NULL,
  multi_scale = FALSE,
  eggnog = TRUE,
  kegg_name_match = "all",
  how = mean,
  eps = 0.01,
  color_list = NULL,
  only_ko = FALSE
)
}
\arguments{
\item{stana}{stana class object}

\item{species}{candidate species}

\item{pathway_id}{KEGG PATHWAY id}

\item{cl}{grouping (named list). If not specified, use stana@cl slot}

\item{multi_scale}{if TRUE, use ``ggh4x`` to plot multiple scales per species}

\item{eggnog}{use eggNOG annotation in stana object}

\item{kegg_name_match}{passed to node_numeric function}

\item{how}{how to concatenate abundance per KO}

\item{eps}{pseudovalue added when claculating log2 fold change}

\item{color_list}{color list if multi_scale=TRUE,
passed to scale_fill_multi()}

\item{only_ko}{only calculates KO and return stana}
}
\description{
plot the KEGG pathway
}
