% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPCA.R
\name{plotPCA}
\alias{plotPCA}
\title{plotPCA}
\usage{
plotPCA(
  stana,
  species,
  cl,
  target = "snps",
  ignore = FALSE,
  pointSize = 5,
  ltype = 2,
  useBlend = FALSE
)
}
\arguments{
\item{stana}{output of midas merge}

\item{species}{candidate species}

\item{cl}{named list of clusters}

\item{target}{snps, genes}

\item{ignore}{ignore the SNVs containing -1}

\item{pointSize}{scatter point size}

\item{ltype}{line type for ellipse}

\item{useBlend}{use ggblend for point}
}
\description{
plot the PCA results from snp or gene table
}
