% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF.R
\name{plotStackedBarPlot}
\alias{plotStackedBarPlot}
\title{plotStackedBarPlot}
\usage{
plotStackedBarPlot(stana, sp, by = "NMF", mat = NULL)
}
\arguments{
\item{stana}{stana boject}

\item{sp}{species}

\item{by}{"NMF", "manual", or "coef"}

\item{mat}{manual matrix}
}
\value{
ggplot
}
\description{
plot the stacked bar plot of NMF results
}
