% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTree.R
\name{plotTree}
\alias{plotTree}
\title{plotTree}
\usage{
plotTree(stana, species = NULL, cl = NULL)
}
\arguments{
\item{stana}{stana object}

\item{species}{species to plot
If NULL, first species is assigned}

\item{cl}{optional, cluster to plot}
}
\description{
infer the tree and plot
}
