% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTree.R
\name{plotTree}
\alias{plotTree}
\title{plotTree}
\usage{
plotTree(
  stana,
  species = NULL,
  cl = NULL,
  model = "F81",
  branch.length = "none"
)
}
\arguments{
\item{stana}{stana object}

\item{species}{species to plot
If NULL, first species in fasta list is assigned}

\item{cl}{optional, cluster to plot}

\item{model}{dist.ml model}

\item{branch.length}{branch length, default to "none", cladogram}
}
\description{
infer the tree and plot
use dist.ml and NJ for inference
}
