% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkEGGNOG.R
\name{summariseAbundance}
\alias{summariseAbundance}
\title{summariseAbundance}
\usage{
summariseAbundance(stana, sp, anno, how = mean)
}
\arguments{
\item{stana}{stana object}

\item{sp}{candidate species}

\item{anno}{annotation tibble}

\item{how}{summarising function, default to mean}
}
\value{
data.frame
}
\description{
Given tibble with `ID` and `value` column,
Obtain data.frame of genes consisting of `ID` within `value`
and summarise them by default `mean`.
}
